+++ NEWS.md
@@ -1,6 +1,8 @@
 # devel
 
 - `fast_anticlustering()` has been re-implemented in C, which is much faster than the previous R implementation, and which does not crash the computer for about N > 250000 elements, as done by [`anticlustering()`](https://github.com/m-Py/anticlust/issues/50)).
+- `fast_anticlustering()` now has an additional argument `exchange_partners`, which can be used to pass custom exchange partners instead of using the default nearest neighbour search
+- new exported function `generate_exchange_partners()`, can be used to address the new argument `exchange_partners` in `fast_anticlustering()`
 
 TODO: Rewrite 2 vignettes due to the rewrite of `fast_anticlustering()`